server/requirements.txt
astroid==2.0.4  
Click==7.0    
Flask==1.0.2  
Flask-SocketIO==3.0.2  
isort==4.3.4  
itsdangerous==1.1.0  
Jinja2==2.10   
lazy-object-proxy==1.3.1  
MarkupSafe==1.0    
mccabe==0.6.1  
python-engineio==2.3.2  
python-socketio==2.0.0  
six==1.11.0 
typed-ast==1.1.0  
Werkzeug==0.14.1 
wrapt==1.10.11
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
WTForms==2.2.1
eventlet==0.24.1
passlib==1.7.1
flask-migrate==2.3.0
flask-jwt-extended==3.13.1